pyproject.toml
[project]
requires-python = ">=3.10"
authors = [
    {name = "Nicolas Suzor", email = "nic@suzor.com"},
]
name = "buttermilk"
version = "0.3.3"
description = ""
readme = "README.md"
dependencies = [
    "shortuuid<2.0.0,>=1.0.13",
    "psutil<6.0.0,>=5.9.5",
    "cloudpathlib<1.0.0,>=0.20.0",
    "requests<3.0.0,>=2.32.3",
    "pathlib<2.0.0,>=1.0.1",
    "pillow>=10.0.0",
    "pyyaml<7.0.0,>=6.0.2",
    "urllib3<3.0.0,>=2.2.2",
    "azureml-core<2.0.0,>=1.57.0",
    "azure-identity<2.0.0,>=1.17.1",
    "hydra-core<2.0.0,>=1.3.2",
    "azure-keyvault<5.0.0,>=4.2.0",
    "pydantic<3.0.0,>=2.8.2",
    "coloredlogs<16.0.0,>=15.0.1",
    "google-cloud-logging<4.0.0,>=3.11.2",
    "json-repair<1.0.0,>=0.28.3",
    "boto3<2.0.0,>=1.35.7",
    "azure-ai-contentsafety<2.0.0,>=1.0.0",
    "azure-cognitiveservices-vision-contentmoderator<2.0.0,>=1.0.0",
    "unidecode<2.0.0,>=1.3.8",
    "huggingface-hub<1.0.0,>=0.26.1",
    "sentencepiece<1.0.0,>=0.2.0",
    "validators<1.0.0,>=0.33.0",
    "replicate<1.0.0,>=0.32.1",
    "keyring>23.4.0",
    "google-crc32c<2.0.0,>=1.6.0",
    "regex<2025.0.0,>=2024.9.11",
    "proto-plus<2.0.0,>=1.25.0",
    "httpx<1.0.0,>=0.27.2",
    "beautifulsoup4<5.0.0,>=4.12.3",
    "rich<14.0.0,>=13.9.4",
    "google-cloud-secret-manager<3.0.0,>=2.22.0",
    "anthropic[vertex]<1.0.0,>=0.42.0",
    "jq<2.0.0,>=1.8.0",
    "google-cloud-aiplatform<2.0.0,>=1.74.0",
    "db-dtypes<2.0.0,>=1.3.1",
    "google-cloud-language<3.0.0,>=2.15.1",
    "gspread<7.0.0,>=6.1.2",
    "anyio<5.0.0,>=4.7.0",
    "slack-bolt<2.0.0,>=1.21.2",
    "google-cloud-bigquery-storage<3.0.0,>=2.27.0",
    "google-generativeai<1.0.0,>=0.7.2",
    "uvicorn<1.0.0,>=0.34.0",
    "google-cloud-pubsub<3.0.0,>=2.27.1",
    "bleach<7.0.0,>=6.2.0",
    "numpy>=1.26.4",
    "promptflow>=0.17.0",
    "promptflow-tracing>=1.17.0",
    "marshmallow==3.23.2",
    "chromadb>=0",
    "readabilipy>=0.3.0",
    "transformers>=4.47.1",
    "evaluate>=0.4.3",
    "fake-useragent>=2.0.3",
    "json5>=0.10.0",
    "newspaper4k>=0.9.3.1",
    "lxml[html-clean]>=5.3.0",
    "weave>=0.51.33",
    "traceloop-sdk>=0.38.5",
    "promptflow-azure>=1.17.2",
    "autogen-ext[azure,gemini,openai]>=0.4.9",
    "autogen-agentchat>=0.4.9.2",
    "autogen-openaiext-client>=0.0.3",
    "sentry-sdk>=2.24.0",
    "flask-restx>=1.3.0",
    "aioconsole>=0.8.1",
    "autogen-core>=0.4.9.2",
    "pdfminer-six>=20250327",
    "pyzotero>=1.6.11",
    "langchain-text-splitters>=0.3.7",
    "langchain>=0.3.21",
    "pytest-anyio>=0.0.0",
    "shiny>=1.4.0",
    "streamlit>=1.45.0",
    "opentelemetry-exporter-gcp-trace>=1.9.0",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0",
    "opentelemetry-exporter-gcp-logging>=1.9.0a0",
]

[dependency-groups]
ml = [
    "torch>=2.5.1,<3.0.0",
    "evaluate<1.0.0,>=0.4.2",
    "einops<1.0.0,>=0.8.0",
    "hf-transfer<1.0.0,>=0.1.8",
    "loky<4.0.0,>=3.4.1",
    "transformers<5.0.0,>=4.44.2",
    "peft<1.0.0,>=0.12.0",
    "azure-ai-ml==1.18.0",
]
dev = [
    "fsspec<2025.0.0,>=2024.6.1",
    "gcsfs<2025.0.0,>=2024.6.1",
    "pytest<9.0.0,>=8.3.2",
    "cmap<1.0.0,>=0.3.0",
    "seaborn<1.0.0,>=0.13.2",
    "scikit-learn<2.0.0,>=1.5.1",
    "pytest-mock<4.0.0,>=3.14.0",
    "pytest-xdist<4.0.0,>=3.6.1",
    "pytest-timeout<3.0.0,>=2.3.1",
    "pytest-cov<6.0.0,>=5.0.0",
    "matplotlib<4.0.0,>=3.9.2",
    "tqdm<5.0.0,>=4.66.5",
    "hydra-colorlog<2.0.0,>=1.2.0",
    "plotly<6.0.0,>=5.24.1",
    "gradio<6.0.0,>=5.4.0",
    "openpyxl<4.0.0,>=3.1.5",
    "google-cloud-bigquery-storage<3.0.0,>=2.27.0",
    "google-cloud-storage<3.0.0,>=2.19.0",
    "google-cloud-pubsub<3.0.0,>=2.27.1",
    "google-cloud-bigquery<4.0.0,>=3.27.0",
    "opencv-python<5.0.0.0,>=4.10.0.84",
    "promptflow-tracing>=1.16.2",
    "anyio>=4.7.0",
    "pandas-stubs>=2.2.3.241126",
    "ipykernel>=6.29.5",
    "buttermilk",
    "pytest-anyio>=0.0.0",
]
ui = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "altair>=5.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["buttermilk"]

[tool.mypy]
python_version = "3.11"  # Specify your target Python version
strict = false
annotations = true      # delayed typing annotations
warn_unused_configs = true # Warn about unused mypy settings
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.uv]
upgrade = true
default-groups = [] # Add the groups you want installed by default

[tool.uv.sources]
buttermilk = { workspace = true }

[tool.black]
line-length = 150

[tool.ruff]
line-length = 150
target-version = "py311"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I", "PL"]
ignore = [
    "RET504",  # Unnecessary variable assignment before return
    "RET503",  # Multiple returns 
    "E501",    # Line too long 
    "TRY003", 
    "D401",     # First line of docstring must be in imperative mood
    "EM101",   # No f-strings or string literals in exceptions
    "RET505",  # Unnecessary else after return
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["buttermilk"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[project.scripts]
bm-chat = "buttermilk.runner.cli:main"
bm-test-agent = "buttermilk.runner.agent_tester:main"

[tool.pytest.ini_options]
markers = [
  "integration: integration tests, don't run by default",
  "scheduled: mark tests to run in scheduled testing",
]

norecursedirs = ["docs", "*.egg-info", "*.egg", "bin", "demo", ".git", "appdir", ".tox", "mediamarkets", "scrapers"]
timeout = 600
default_async_timeout = 300
addopts = [ 
    "--capture=fd", 
    "--continue-on-collection-errors",
    "--durations=10", 
    "-n 24", 
    "-v", # show name of tests 
    "-rfEX", # show extra summary info, including for failed, eXpected failures, and Errors
    "-s",
    "--tb=auto",
    # "--show-capture=no",
    # "--disable-warnings",
    "-m not integration",
]

#  "--timeout_method=thread", 
# "--with-integration"
# 

filterwarnings = [
    "ignore:Deprecated call to `pkg_resources.declare_namespace:DeprecationWarning",
    "ignore:Passing field metadata as keyword arguments is deprecated",
    "ignore:.*grpc_wait_for_shutdown_with_timeout.*:Warning",
]
